refactor(config): refactor config page axios api
