+++ src/common/api/branch.ts
@@ -1,47 +1,43 @@
-import { UserConfigInfoModel } from '@/common/model'
 import axios from '@/utils/axios'
-
-const README = `
-# Welcome to use PicX
-
-[PicX](https://github.com/XPoet/picx) is a simple and powerful image hosting tool. It supports image hosting services via GitHub repository.
-
-PicX is completely open source, and you can use it for free.
-
-If you like it, please give it a star on [GitHub](https://github.com/XPoet/picx).
-        `
+import { UserConfigInfoModel } from '@/common/model'
 
 /**
- * 初始化一个空仓库
- * @param userConfigInfo
+ * 获取分支信息列表
+ * @param owner
+ * @param repo
+ * @param callback
  */
-export async function initRepo(userConfigInfo: UserConfigInfoModel) {
-  const { owner, selectedRepo: repo, selectedBranch: branch } = userConfigInfo
-
-  const initRepoLoading = ElLoading.service({
-    text: '正在初始化仓库...'
-  })
-
-  // GitHub Git database API 不支持在空仓库上操作，需要先初始化空仓库
-  // 仓库为空时，新建一个 README 文件来初始化仓库
-  const res = await axios.put(`/repos/${owner}/${repo}/contents/README.md`, {
-    message: 'Init repo via PicX(https://github.com/XPoet/picx)',
-    branch,
-    content: window.btoa(README)
-  })
-
-  if (res?.status === 201) {
-    initRepoLoading.close()
-  } else {
-    ElMessage.error('仓库初始化失败')
-  }
+export const getBranchInfoList = <T extends (...args: any) => void>(
+  owner: string,
+  repo: string,
+  callback: T
+) => {
+  axios
+    .get(`/repos/${owner}/${repo}/branches`)
+    .then(async (res: any) => {
+      console.log('getBranchList >> ', res)
+      if (res && res.status === 200) {
+        callback(res.data)
+      } else {
+        callback(null)
+      }
+    })
+    .catch(() => {
+      callback(null)
+    })
 }
 
-export async function initBranch(
+/**
+ * 初始化空分支（未实现，待完善）
+ * @param userConfigInfo
+ * @param branch
+ * @param callback
+ */
+export const initBranch = async (
   userConfigInfo: UserConfigInfoModel,
   branch: string,
   callback: any
-) {
+) => {
   const { owner, selectedRepo: repo, branchList } = userConfigInfo
 
   const initLoading = ElLoading.service({
