+++ src/utils/index.ts
@@ -3,4 +3,3 @@ export * from './compress-image'
 export * from './image-link-utils'
 export * from './image-utils'
 export * from './file-utils'
-export * from './init-repo'
